+++ src/features/gathering/model/createGatheringSchema.ts
@@ -1,0 +1,44 @@
+import { z } from 'zod';
+
+export const createGatheringSchema = z
+  .object({
+    name: z.string().min(1, 'form.errors.nameRequired'),
+    location: z.enum(['건대입구', '을지로3가', '신림', '홍대입구'], {
+      required_error: 'form.errors.locationRequired',
+    }),
+    type: z.enum(['OFFICE_STRETCHING', 'MINDFULNESS', 'WORKATION'], {
+      required_error: 'form.errors.typeRequired',
+    }),
+    dateTime: z.date({
+      required_error: 'form.errors.dateTimeRequired',
+    }),
+    registrationEnd: z.date({
+      required_error: 'form.errors.registrationEndRequired',
+    }),
+    capacity: z.coerce.number().min(5, 'form.errors.capacityMin'),
+    image: z
+      .any()
+      .optional()
+      .refine(
+        (file) =>
+          !file ||
+          (typeof File !== 'undefined' && file instanceof File && file.type.startsWith('image/')),
+        {
+          message: 'form.errors.imageType',
+          path: ['image'],
+        },
+      ),
+  })
+  .refine(
+    (data) => !data.registrationEnd || (data.dateTime && data.registrationEnd < data.dateTime),
+    {
+      message: 'form.errors.registrationEndInvalid',
+      path: ['registrationEnd'],
+    },
+  )
+  .refine((data) => data.dateTime && data.dateTime > new Date(), {
+    message: 'form.errors.dateTimePast',
+    path: ['dateTime'],
+  });
+
+export type CreateGatheringFormValues = z.infer<typeof createGatheringSchema>;
