+++ src/features/gathering/model/index.ts
@@ -1,0 +1,2 @@
+export { createGatheringSchema } from './createGatheringSchema';
+export type { CreateGatheringFormValues } from './createGatheringSchema';
